+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `models.TextChoices` refactored for better code readability
 - Models refactored
 - Views refactored
+- Switched to `allianceauth-app-utils` for logging
 
 
 ## [1.2.0] - 2022-01-12